Remove dependencies and code for displaying plots
SciJava Plot should not depend on any swing dependent libraries like
JFreeChart or scijava-ui-swing.

The code for displaying plots is therefor moved to scijava-ui-swing
* This includes the converters:
 * XYPlot -> JFreeChart
 * CategoryChart -> JFreeChart
* An io plugin to save plots as SVG,
* Displays for JFreeChart

The code for converting plots to ImagePlus and Img is moved to
imagej-legacy as it depends on ImagePlus. A converter from
plot to java.awt.image.BufferedImage will be added to scijava-ui-swing.